+++ src/client.cpp
@@ -12,120 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <boost/thread/thread.hpp>
 #include <protolink/client.hpp>
-
-namespace protolink
-{
-namespace udp_protocol
-{
-Publisher::Publisher(
-  boost::asio::io_service & io_service, const std::string & ip_address, const uint16_t port,
-  const uint16_t from_port, const rclcpp::Logger & logger)
-: endpoint(boost::asio::ip::udp::endpoint(
-    boost::asio::ip::udp::endpoint(boost::asio::ip::address::from_string(ip_address), port))),
-  logger(logger),
-  sock_(io_service, boost::asio::ip::udp::endpoint(boost::asio::ip::udp::v4(), from_port))
-{
-}
-
-void Publisher::sendEncodedText(const std::string & encoded_text)
-{
-  sock_.send_to(boost::asio::buffer(encoded_text), endpoint);
-}
-}  // namespace udp_protocol
-
-namespace serial_protocol
-{
-Publisher::Publisher(
-  boost::asio::io_service & io_service, const std::string & device_file, const uint16_t baud_rate,
-  const rclcpp::Logger & logger)
-: logger(logger), serial_(io_service, device_file)
-{
-  serial_.set_option(boost::asio::serial_port_base::baud_rate(baud_rate));
-}
-
-void Publisher::sendEncodedText(const std::string & encoded_text)
-{
-  boost::asio::write(serial_, boost::asio::buffer(encoded_text));
-}
-
-Subscriber::Subscriber(
-  boost::asio::io_service & io_service, const std::string & device_file, const uint16_t baud_rate,
-  const rclcpp::Logger & logger)
-: logger(logger), serial_(io_service, device_file)
-{
-  serial_.set_option(boost::asio::serial_port_base::baud_rate(baud_rate));
-  constexpr int receive_size = 8;
-  boost::asio::async_read(
-    serial_, boost::asio::buffer(receive_data_), boost::asio::transfer_exactly(receive_size),
-    boost::bind(
-      &Subscriber::handler, this, boost::asio::placeholders::error,
-      boost::asio::placeholders::bytes_transferred));
-}
-
-void Subscriber::handler(const boost::system::error_code & /*error*/, size_t /*bytes_transferred*/)
-{
-  size_t len = 64;
-  std::string s(reinterpret_cast<char const *>(receive_data_), len);
-  protolink__std_msgs__String::std_msgs__String proto;
-  proto.SerializeToString(&s);
-}
-}  // namespace serial_protocol
-
-// namespace mqtt_protocol
-// {
-// Publisher::Publisher(
-//   const std::string & server_address, const std::string & client_id, const std::string & topic,
-//   const int qos, const rclcpp::Logger & logger)
-// : topic(topic),
-//   qos(qos),
-//   logger(logger),
-//   client_impl_(server_address, client_id, "/tmp/mqtt"),
-//   connection_thread_running_(true)
-// {
-//   client_impl_.set_callback(callback_);
-//   auto connect_options = mqtt::connect_options_builder()
-//                            .clean_session()
-//                            .will(mqtt::message(topic, LWT_PAYLOAD, strlen(LWT_PAYLOAD), qos, false))
-//                            .finalize();
-//   connect_token_ = client_impl_.connect(connect_options);
-//   connection_thread_ = std::thread([this]() {
-//     const auto connect = [this]() {
-//       try {
-//         if (!client_impl_.is_connected()) {
-//           mqtt::token_ptr conntok = client_impl_.connect();
-//           conntok->wait();
-//         }
-//       } catch (const mqtt::exception & exc) {
-//         RCLCPP_ERROR_STREAM(this->logger, exc.what());
-//       }
-//     };
-//     while (connection_thread_running_) {
-//       connect();
-//       std::this_thread::sleep_for(std::chrono::seconds(1));
-//     }
-//   });
-// }
-
-// Publisher::~Publisher()
-// {
-//   connection_thread_running_ = false;
-//   connection_thread_.join();
-// }
-
-// void Publisher::sendEncodedText(const std::string & encoded_text)
-// {
-//   if (client_impl_.is_connected()) {
-//     client_impl_.publish(topic, encoded_text.c_str(), encoded_text.size(), qos, false);
-//   }
-// }
-
-// Subscriber::Subscriber(
-//   const std::string & server_address, const std::string & client_id, const std::string & topic,
-//   const int qos, const rclcpp::Logger & logger)
-// : topic(topic), qos(qos), logger(logger), client_impl_(server_address, client_id)
-// {
-// }
-// }  // namespace mqtt_protocol
-}  // namespace protolink
